+++ react/components/Preview/Box.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
-import ContentLoader from './ContentLoader'
+import { ContentLoader, Rect } from './ContentLoader'
 
 interface Props {
-  width: number
-  height: number
+  width: number | string
+  height: number | string
 }
 
 const Box = ({ width, height }: Props) => (
   <ContentLoader width={width} height={height}>
-    <rect x="0" y="0" rx="5" ry="5" width={width} height={height} />
+    <Rect x={0} y={0} width={width} height={height} />
   </ContentLoader>
 )