+++ react/components/Preview/Spinner.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
 interface Props {
-  width: number
-  height: number
+  width: number | string
+  height: number | string
 }
 
 const Spinner = ({ width, height }: Props) => (